rewroted routing for folder module
