+++ src/app/modules/app-routing.module.ts
@@ -13,10 +13,6 @@ const routes: Routes = [
     path: 'auth/registration',
     loadChildren: () => import ('./auth/auth.module').then(a =>    a.AuthModule),
   },
-  {
-    path: 'home',
-    loadChildren: () => import('./mentor/mentor-panel.module').then(m => m.MentorPanelModule),
-  },
 ];
 
 @NgModule({
