+++ src/app/modules/folders/components/folder-items-list/folder-items-list.component.html
@@ -1,2 +1,2 @@
 <p>folder-items-list works!</p>
-<router-outlet></router-outlet>
+{{ id }}
