+++ src/app/modules/folders/folder.routing.ts
@@ -1,13 +1,15 @@
 import { Routes } from '@angular/router';
+import { FolderView } from './views/folder-view/folder.view';
 
 export const folderRoutes: Routes = [
   {
-    path: '',
+    path: 'home/:id',
+    component: FolderView,
     children: [
       {
-        path: '/new-question',
+        path: 'new-question',
         loadChildren: () => import('../questions/question.module').then(q => q.QuestionModule),
       },
     ],
-  }
+  },
 ];
