+++ src/app/modules/folders/views/folder-view/folder.view.html
@@ -1,0 +1,22 @@
+<div>
+  <nav>sub menu here</nav>
+  <app-folder-items-list></app-folder-items-list>
+
+  <nav>
+    <ul>
+      <li><a routerLink="new-question">
+        <img src="../../../../../assets/add-question-icon.png" alt="link for add new question">
+      </a></li>
+      <!-- <li>
+        <a routerLink="/home/new-test">
+          <img class="add-icon" src="../../../../../../assets/add-folder-icon.png" alt="add folder icon">
+        </a>
+      </li>
+      <li>
+        <a routerLink="/home/new-test">
+          <img class="add-icon" src="../../../../../../assets/add-test-icon.png" alt="add test icon">
+        </a>
+      </li> -->
+    </ul>
+  </nav>
+</div>
