--- src/app/modules/mentor/folders/components/folder-items-list/folder-items-list.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-folder-items-list',
-  templateUrl: './folder-items-list.component.html',
-  styleUrls: ['./folder-items-list.component.sass']
-})
-export class FolderItemsListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
