--- src/app/modules/mentor/folders/views/folder-view/folder.view.html
@@ -1,4 +1,0 @@
-<div>
-  <p>folder-view works!</p>
-  <p>mentor-panel-views works! - old</p>
-</div>
