--- src/app/modules/mentor/mentor-view/mentor.view.html
@@ -1,13 +1,0 @@
-<router-outlet name="mentor"></router-outlet>
-<p>work mentor</p>
-<nav>
-  <a routerLink="/home/new-question">
-    <img class="add-icon" src="../../../../../../assets/add-question-icon.png" alt="add question icon">
-  </a>
-  <a routerLink="/home/new-test">
-    <img class="add-icon" src="../../../../../../assets/add-test-icon.png" alt="add test icon">
-  </a>
-  <a routerLink="/home/new-test">
-    <img class="add-icon" src="../../../../../../assets/add-folder-icon.png" alt="add folder icon">
-  </a>
-</nav>
