--- src/app/modules/mentor/mentor-view/mentor.view.sass
@@ -1,2 +1,0 @@
-.add-icon
-  height: 5rem
