--- src/app/modules/mentor/mentor-view/mentor.view.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-mentor',
-  templateUrl: './mentor.view.html',
-  styleUrls: ['./mentor.view.sass']
-})
-export class MentorView implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
