+++ src/app/modules/questions/question.routing.ts
@@ -1,10 +1,10 @@
 import { Routes } from '@angular/router';
-import { MentorGuard } from '../mentor.guard';
+import { MentorGuard } from '../../utils/mentor.guard';
 import { CreatingQuestionView } from './views/creating-question-view/creating-question.view';
 
 export const questionRoutes: Routes = [
   {
-    path: 'home/new-question',
+    path: 'home/:id/new-question',
     component: CreatingQuestionView,
     canActivate: [MentorGuard],
   },
